requirements.txt
python-dotenv==0.9.1
requests==2.20.1
terminaltables==3.1.0

